backend-ai/requirements.txt
langchain==0.2.16
langchain-groq==0.1.9
langchain-core==0.2.38
groq==0.9.0
supabase==2.7.4
python-dotenv==1.0.1
fastapi==0.112.2
uvicorn==0.30.6
